requirements.txt
python-docx
python-dotenv

# Graphiti
graphiti-core

# LangChain
unstructured
langchain_experimental
langchain_community
langchain_openai